fix(scale): resolve WiFi scale via mDNS A-record query on Android, not NsdManager
The NsdManager approach shipped in #1247 did not find the scale: it used
DNS-SD service discovery (discoverServices "_http._tcp"), but the Half
Decent Scale publishes a plain "hds.local" mDNS HOSTNAME, not a DNS-SD
service record — so discovery returned nothing every time. Device log
showed "NSD probe timed out" / "NSD lookup failed" on every attempt while
MqttClient resolved its own ".local" broker fine in the same session.

Switch the Android path to a direct mDNS A-record query — the exact
mechanism MqttClient already uses successfully via mjansson/mdns.

- Extract MqttClient's private resolveMdns() into a shared
  src/network/mdnsresolver.{h,cpp} (MdnsResolver::resolveHostname). This
  TU now owns the single MDNS_IMPLEMENTATION definition for the header-only
  mjansson library; mqttclient.cpp calls into it.
- WifiScaleDiscovery's Android branch runs MdnsResolver::resolveHostname on
  a QThreadPool worker and posts the result back with the existing
  generation-counter guard. cancelInFlight() simplifies to a generation
  bump (no NsdManager listener/JNI to tear down).
- Delete WifiScaleNsdHelper.java (NsdManager helper no longer used).

Multicast reception relies on the process-wide WifiManager.MulticastLock
ShotServer holds for the app lifetime — same arrangement MqttClient
depends on.

macOS build: 0 errors / 0 warnings (mdnsresolver compiles to a no-op stub
off Android; WifiScaleDiscovery uses QHostInfo there). Android compile
verified separately via CI.
