+++ src/network/wifiscalediscovery.h
@@ -14,10 +14,10 @@ class QTimer;
  * followed by probeFinished(). On timeout / NotFound emits only
  * probeFinished(). Calling probe() while a lookup is in flight starts a
  * new lookup; the previous lookup's result will be dropped. On non-Android
- * the previous QHostInfo lookup is also synchronously aborted; on Android
- * the previous NSD discovery is torn down via WifiScaleNsdHelper.cancelDiscovery()
- * (the Java worker thread is unblocked but cannot be synchronously joined —
- * the result, when it eventually arrives, is dropped via a generation check).
+ * the previous QHostInfo lookup is synchronously aborted. On Android the
+ * lookup runs on a worker thread via MdnsResolver (a direct mDNS A-record
+ * query, since the OS resolver doesn't speak mDNS); the worker cannot be
+ * synchronously joined, so its late result is dropped via a generation check.
  */
 class WifiScaleDiscovery : public QObject {
     Q_OBJECT
